BLEmonitor on integration level, BLEparser stays on platform
