+++ custom_components/ble_monitor/__init__.py
@@ -1,5 +1,8 @@
 """Passive BLE monitor integration."""
+import asyncio
 import logging
+import queue
+from threading import Thread
 import voluptuous as vol
 from homeassistant.helpers import config_validation as cv
 
@@ -9,9 +12,15 @@ from homeassistant.const import (
     CONF_MAC,
     CONF_NAME,
     CONF_TEMPERATURE_UNIT,
+    EVENT_HOMEASSISTANT_STOP,
 )
 
 from homeassistant.helpers import discovery
+
+# It was decided to temporarily include this file in the integration bundle
+# until the issue with checking the adapter's capabilities is resolved in the official aioblescan repo
+# see https://github.com/frawau/aioblescan/pull/30, thanks to @vicamo
+from . import aioblescan_ext as aiobs
 
 from .const import (
     DEFAULT_ROUNDING,
@@ -87,6 +96,133 @@ CONFIG_SCHEMA = vol.Schema(
 
 def setup(hass, config):
     """Set up integration."""
-    hass.data[DOMAIN] = config[DOMAIN]
+    blemonitor = BLEmonitor(config[DOMAIN])
+    hass.bus.listen(EVENT_HOMEASSISTANT_STOP, blemonitor.shutdown_handler)
+    blemonitor.start()
+    hass.data[DOMAIN] = blemonitor
     discovery.load_platform(hass, "sensor", DOMAIN, {}, config)
     return True
+
+
+class BLEmonitor:
+    """BLE scanner."""
+
+    def __init__(self, config):
+        """Init."""
+        self.dataqueue = queue.Queue()
+        self.config = config
+        self.dumpthread = None
+
+    def shutdown_handler(self, event):
+        """Run homeassistant_stop event handler."""
+        _LOGGER.debug("Shutdown event fired: %s", event)
+        self.stop()
+
+    def start(self):
+        """Start receiving broadcasts."""
+        _LOGGER.debug("Spawning HCIdump thread")
+        self.dumpthread = HCIdump(
+            config=self.config,
+            dataqueue=self.dataqueue,
+        )
+        self.dumpthread.start()
+
+    def stop(self):
+        """Stop HCIdump thread(s)."""
+        self.dataqueue.put(None)
+        result = True
+        if self.dumpthread is None:
+            _LOGGER.debug("BLE monitor stopped")
+            return True
+        if self.dumpthread.is_alive():
+            self.dumpthread.join()
+            if self.dumpthread.is_alive():
+                result = False
+                _LOGGER.error(
+                    "Waiting for the HCIdump thread to finish took too long! (>10s)"
+                )
+        _LOGGER.debug("BLE monitor stopped")
+        return result
+
+    def restart(self):
+        """Restart scanning"""
+        if self.dumpthread.is_alive():
+            self.dumpthread.restart()
+        else:
+            self.start()
+
+
+class HCIdump(Thread):
+    """Mimic deprecated hcidump tool."""
+
+    def __init__(self, config, dataqueue):
+        """Initiate HCIdump thread."""
+        Thread.__init__(self)
+        _LOGGER.debug("HCIdump thread: Init")
+        self._interfaces = config[CONF_HCI_INTERFACE]
+        self._active = int(config[CONF_ACTIVE_SCAN] is True)
+        self.dataqueue = dataqueue
+        self._event_loop = None
+        self._joining = False
+
+    def process_hci_events(self, data):
+        """Collect HCI events."""
+        self.dataqueue.put(data)
+
+    def run(self):
+        """Run HCIdump thread."""
+        while True:
+            _LOGGER.debug("HCIdump thread: Run")
+            mysocket = {}
+            fac = {}
+            conn = {}
+            btctrl = {}
+            if self._event_loop is None:
+                self._event_loop = asyncio.new_event_loop()
+                asyncio.set_event_loop(self._event_loop)
+            for hci in self._interfaces:
+                try:
+                    mysocket[hci] = aiobs.create_bt_socket(hci)
+                except OSError as error:
+                    _LOGGER.error("HCIdump thread: OS error (hci%i): %s", hci, error)
+                else:
+                    fac[hci] = getattr(self._event_loop, "_create_connection_transport")(
+                        mysocket[hci], aiobs.BLEScanRequester, None, None
+                    )
+                    conn[hci], btctrl[hci] = self._event_loop.run_until_complete(fac[hci])
+                    _LOGGER.debug("HCIdump thread: connected to hci%i", hci)
+                    btctrl[hci].process = self.process_hci_events
+                    self._event_loop.run_until_complete(btctrl[hci].send_scan_request(self._active))
+            _LOGGER.debug("HCIdump thread: start main event_loop")
+            try:
+                self._event_loop.run_forever()
+            finally:
+                _LOGGER.debug("HCIdump thread: main event_loop stopped, finishing")
+                for hci in self._interfaces:
+                    self._event_loop.run_until_complete(btctrl[hci].stop_scan_request())
+                    conn[hci].close()
+                self._event_loop.run_until_complete(asyncio.sleep(0))
+            if self._joining is True:
+                break
+            _LOGGER.debug("HCIdump thread: Scanning will be restarted")
+        self._event_loop.close()
+        _LOGGER.debug("HCIdump thread: Run finished")
+
+    def join(self, timeout=10):
+        """Join HCIdump thread."""
+        _LOGGER.debug("HCIdump thread: joining")
+        self._joining = True
+        try:
+            self._event_loop.call_soon_threadsafe(self._event_loop.stop)
+        except AttributeError as error:
+            _LOGGER.debug("%s", error)
+        finally:
+            Thread.join(self, timeout)
+            _LOGGER.debug("HCIdump thread: joined")
+
+    def restart(self):
+        """Restarting scanner"""
+        try:
+            self._event_loop.call_soon_threadsafe(self._event_loop.stop)
+        except AttributeError as error:
+            _LOGGER.debug("%s", error)
